Rename ReflectionTestClass.php
